+++ crates/project/src/resolver/mod.rs
@@ -922,6 +922,9 @@ enum SourceVersionError {
 
 #[cfg(test)]
 mod tests {
+    use std::fs::{self, create_dir_all};
+    use utils::tempdir;
+
     use super::*;
 
     #[test]
@@ -1007,5 +1010,30 @@ src/Dapp.t.sol >=0.6.6
 ",
             String::from_utf8(out).unwrap()
         );
+    }
+
+    #[cfg(target_os = "linux")]
+    #[test]
+    fn can_read_different_case() {
+        use crate::resolver::parse::SolData;
+
+        let tmp_dir = tempdir("out").unwrap();
+        let path = tmp_dir.path().join("forge-std");
+        create_dir_all(&path).unwrap();
+        let existing = path.join("Test.sol");
+        let non_existing = path.join("test.sol");
+        fs::write(
+            existing,
+            "
+pragma solidity ^0.8.10;
+contract A {}
+        ",
+        )
+        .unwrap();
+
+        assert!(!non_existing.exists());
+
+        let found = crate::resolver::Node::<SolData>::read(&non_existing).unwrap_err();
+        matches!(found, SolcError::ResolveCaseSensitiveFileName { .. });
     }
 }
